feat: redesign batch actions
