+++ packages/web-app-files/src/composables/index.ts
@@ -1,2 +1,3 @@
 export * from './actions'
 export * from './resourcesViewDefaults'
+export * from './useFileUpload'
