+++ packages/web-pkg/src/components/BatchActions.vue
@@ -2,7 +2,7 @@
   <div>
     <oc-list
       id="oc-appbar-batch-actions"
-      class="block xl:flex xl:items-center"
+      class="flex items-center gap-1"
       :class="{
         'oc-appbar-batch-actions-squashed [&_.oc-files-context-action-label]:hidden':
           limitedScreenSpace
@@ -14,9 +14,10 @@
         :action="action"
         :action-options="actionOptions"
         appearance="raw"
-        class="batch-actions mr-2 float-left [&_.action-menu-item]:p-2 [&_.action-menu-item]:gap-1"
+        class="batch-actions float-left [&_.action-menu-item]:p-2 [&_.action-menu-item]:gap-1 [&_button]:py-1!"
         :shortcut-hint="false"
         :show-tooltip="limitedScreenSpace"
+        :size="limitedScreenSpace ? 'medium' : 'small'"
       />
     </oc-list>
   </div>
